+++ include/stackvm/vm.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace stackvm {
+
+class VM {
+public:
+    VM();
+    ~VM();
+
+    void run();
+};
+
+}